lib/streamaggr: reduce memory usage during deduplication
The change introduces a new deduplication logic instead of `last` aggregation.
Changes to deduplication are the following:
* deduplication logic now keeps the sample with latest `timestamp`
or biggest `value` if timestamps are identical. This logic is aligned with
deduplication logic in vmstorage and vmselect.
* during deduplication, time series are encoded with a custom data structure
to reduce memory footprint when deduplicating a big number of unique time series.
* deduplication can be configured individually per each aggregator. It also means that
deduplication happens individually for each aggregator.

Considerartions:
1. Deduplication happens once per configured interval
2. During the flush, deduplicator tries to push collected samples to aggregator as fast as possible
3. It could happen that flush will take more time than configured deduplication interval, which would
result in compromised aggregation correctness.
4. New metrics `vmagent_streamaggr_dedup_duration_seconds` and `vmagent_streamaggr_dedup_duration_exceeds_aggregation_interval_total` were added to track the duration
of flush interval and number of times when interval was exceeded.
Alerting rule `increase(vmagent_streamaggr_dedup_duration_exceeds_aggregation_interval_total)>0` could help
to detect events when deduplication can't keep up with the load. In this case, user should either adjust resources
and configuration of aggregating component, or reduce the load.

---------------

This change is based on https://github.com/VictoriaMetrics/VictoriaMetrics/pull/5803

Signed-off-by: hagen1778 <roman@victoriametrics.com>